chore: global rename from se2game to saboteur to match shared models
