+++ app/app/src/main/java/com/aau/saboteur/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.aau.se2game.ui.theme
+package com.aau.saboteur.ui.theme
 
 import android.app.Activity
 import android.os.Build
